created a search bar
Created a search bar that redirects to google.

Also added bootstrap to the mix, you know to help with columns and everything else.

I plan on having a twitch integration that shows you who you are following and if they are live.

Your top sites will go on the right side and will list vertically.

I will also add the obligatory to-do list. I thought it would be a nice add to make sure your "need to dos" are always in your face when you click new tab. :P